unipitt: use Configuration to pass command line arguments
Previously, unipitt.Handler got the values of command line arguments
as parameters as well as was responsible for reading the configuration
file. This made it a bit difficult to add new and new parameters (such
as MQTT authentication data, or as planned a prefix for topic names).

To remediate this, this patch gets rid of the separate parameters to
NewHandler() and relies on a Configuration instance to carry arguments
to the constructor.

This allocates reponsibility a bit better, NewHandler does not need to care
what is in the configuration and what was passed as command line arguments.

This simply becomes a user interface issue.

Signed-off-by: Balazs Scheidler <bazsi77@gmail.com>